requirements.txt
bcrypt==5.0.0
bleach==6.3.0
blinker==1.9.0
cachelib==0.13.0
click==8.3.3
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.4.0
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.3.0
greenlet==3.4.0
gunicorn==25.3.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
Markdown==3.10.2
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
Pygments==2.20.0
python-dotenv==1.2.2
python-slugify==8.0.4
SQLAlchemy==2.0.49
text-unidecode==1.3
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.8
wrapt==2.1.2
WTForms==3.2.1
